Click me to see full commit message
Firstly, sorry for one big commit, I got carried away and this is my
first time working on a project with someone else...
All of these changes are for creating web scrapers which are uniform in
structure and output. Design patterns were followed (to the best of
my ability). There is a helper text file because I was such a fool in
making such a large commit.